django/requirements.txt
Django==3.2
django-environ==0.10.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.5

djangorestframework
djangorestframework-simplejwt
